feat.  Feign

+++ module-service/src/main/java/org/example/dto/FeignClientResponseDto.java
@@ -2,11 +2,9 @@ package org.example.dto;
 
 import lombok.Getter;
 
-import javax.validation.constraints.NotNull;
-
 @Getter
 public class FeignClientResponseDto {
-    private Integer userId;
+    private Long userId;
     private Long id;
     private String title;
     private String body;
